Pull classes to package level
