+++ server/src/main/java/io/spine/server/event/enrich/EnrichmentFunction.java
@@ -51,7 +51,7 @@ abstract class EnrichmentFunction<S, T, C extends Message> {
      * We are having the generified class to be able to bound the types of messages and the
      * translation function when building the {@link Enricher}.
      *
-     * @see Enricher.Builder#add(Class, Class, java.util.function.BiFunction)
+     * @see Builder#add(Class, Class, java.util.function.BiFunction)
      */
 
     private final Class<S> sourceClass;
