+++ server/src/test/java/io/spine/server/event/given/ReferenceValidatorTestEnv.java
@@ -27,6 +27,7 @@ import io.spine.core.EventContext;
 import io.spine.core.EventId;
 import io.spine.core.UserId;
 import io.spine.people.PersonName;
+import io.spine.server.event.enrich.Builder;
 import io.spine.server.event.enrich.Enricher;
 import io.spine.test.event.ProjectId;
 import io.spine.testing.core.given.GivenUserId;
@@ -77,7 +78,7 @@ public class ReferenceValidatorTestEnv {
 
         /** Creates a new enricher with all required enrichment functions set. */
         public static Enricher newEventEnricher() {
-            Enricher.Builder builder = Enricher
+            Builder builder = Enricher
                     .newBuilder()
                     .add(ProjectId.class, String.class, new GetProjectName())
                     .add(ProjectId.class, UserId.class, new GetProjectOwnerId())
